pyproject.toml
[tool.poetry]
name = "emakler-etfs"
version = "0.1.0"
description = ""
authors = ["Paweł Dąbrowski"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.9"
camelot-py = {extras = ["base"], version = "^0.10.1"}
ipykernel = "^6.6.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
